メッシュ作っても大丈夫

+++ src/graphics/graphics.cpp
@@ -136,10 +136,10 @@ void initialize(const Config &config) {
 
 void createMesh(
 	const char *id,
-	const unsigned int vertexCount,
+	const uint32_t vertexCount,
 	const float *vertices,
-	const unsigned int indexCount,
-	const float *indices
+	const uint32_t indexCount,
+	const uint32_t *indices
 ) {
 	mesh::createMesh(g_physicalDevice.getMemoryProperties(), g_device, id, vertexCount, vertices, indexCount, indices);
 }
