+++ src/graphics/graphics.hpp
@@ -11,10 +11,10 @@ void initialize(const Config &config);
 
 void createMesh(
 	const char *id,
-	const unsigned int vertexCount,
+	const uint32_t vertexCount,
 	const float *vertices,
-	const unsigned int indexCount,
-	const float *indices
+	const uint32_t indexCount,
+	const uint32_t *indices
 );
 
 void beginRender();
