+++ src/graphics/mesh.cpp
@@ -20,10 +20,10 @@ void createMesh(
 	const vk::PhysicalDeviceMemoryProperties &physicalDeviceMemoryProps,
 	const vk::Device &device,
 	const char *id,
-	const unsigned int vertexCount,
+	const uint32_t vertexCount,
 	const float *vertices,
-	const unsigned int indexCount,
-	const float *indices
+	const uint32_t indexCount,
+	const uint32_t *indices
 ) {
 	const auto vertexBuffer = device.createBuffer(
 		vk::BufferCreateInfo()
