+++ src/graphics/mesh.hpp
@@ -9,10 +9,10 @@ void createMesh(
 	const vk::PhysicalDeviceMemoryProperties &physicalDeviceMemoryProps,
 	const vk::Device &device,
 	const char *id,
-	const unsigned int vertexCount,
+	const uint32_t vertexCount,
 	const float *vertices,
-	const unsigned int indexCount,
-	const float *indices
+	const uint32_t indexCount,
+	const uint32_t *indices
 );
 
 void terminate(const vk::Device &device);
